+++ src/pyconmech/frame_analysis/io_base.py
@@ -1,15 +1,74 @@
+import os
+import json
 import datetime
+import warnings
+from collections import defaultdict
+
+'''length scale conversion to meter'''
+LENGTH_SCALE_CONVERSION = {
+    'millimeter': 1e-3,
+    'meter': 1.0,
+}
+
+#############################################
+
+def parse_point(json_point, scale=1.0):
+    return [scale*json_point[0], scale*json_point[1], scale*json_point[2]]
+
+def parse_nodes(node_data, scale=1.0):
+    nodes = []
+    for n in node_data:
+        n['point'] = parse_point(n['point'], scale=scale)
+        nodes.append(Node.from_data(n))
+    return nodes
+
+def extract_model_name_from_path(file_path):
+    forename = file_path.split('.json')[0]
+    return forename.split(os.sep)[-1]
+
+#############################################
 
 class Model(object):
-    def __init__(self, nodes, elements, supports, joints, materials, crosssecs, model_name=None):
+    def __init__(self, nodes, elements, supports, joints, materials, crosssecs, unit='meter', model_name=None):
         self.model_name = model_name
         self.generate_time = str(datetime.datetime.now())
+        assert unit in LENGTH_SCALE_CONVERSION
+        self.unit = unit
+        scale = LENGTH_SCALE_CONVERSION[unit]
+        if scale != 1.0:
+            warnings.warn('Model unit {}, scaled by {} and converted to Meter unit.'.format(unit, scale))
+        for node in nodes:
+            node.point = [scale*cp for cp in node.point]
+
         self.nodes = nodes
         self.elements = elements
-        self.supports = supports
-        self.joints = joints
-        self.crosssecs = crosssecs
-        self.materials = materials
+
+        # turn lists into dicts
+        self.supports = {}
+        for support in supports:
+            self.supports[support.node_ind] = support
+
+        self.joints = {}
+        if joints is not None:
+            for joint in joints:
+                for e_tag in joint.elem_tags:
+                    if e_tag in self.joints:
+                        warnings.warn('Multiple joints assigned to the same element tag |{}|!'.format(e_tag))
+                    self.joints[e_tag] = joint
+
+        self.materials = {}
+        for mat in materials:
+            for e_tag in mat.elem_tags:
+                if e_tag in self.materials:
+                    warnings.warn('Multiple materials assigned to the same element tag |{}|!'.format(e_tag))
+                self.materials[e_tag] = mat
+
+        self.crosssecs = {}
+        for cs in crosssecs:
+            for e_tag in cs.elem_tags:
+                if e_tag in self.crosssecs:
+                    warnings.warn('Multiple materials assigned to the same element tag |{}|!'.format(e_tag))
+                self.crosssecs[e_tag] = cs
 
     @property
     def node_num(self):
@@ -20,15 +79,94 @@ class Model(object):
         return len(self.elements)
 
     @classmethod
-    def from_data(cls, data):
-        from pyconmech.frame_analysis.frame_file_io import read_frame_data
-        nodes, elements, supports, joints, materials, crosssecs, model_name, unit = read_frame_data(data)
+    def from_json(cls, file_path, verbose=False):
+        assert os.path.exists(file_path) and "json file path does not exist!"
+        with open(file_path, 'r') as f:
+            json_data = json.loads(f.read())
+        if 'model_name' not in json_data:
+            json_data['model_name']  = extract_model_name_from_path(file_path)
+        return cls.from_data(json_data, verbose)
+
+    @classmethod
+    def from_data(cls, data, verbose=False):
+        model_name = data['model_name']
+        unit = data['unit']
+        # length scale for vertex positions
+        scale = LENGTH_SCALE_CONVERSION[unit]
+        unit = 'meter'
+        nodes = parse_nodes(data['nodes'], scale=scale)
+        elements = [Element.from_data(e) for e in data['elements']]
+        element_inds_from_tag = defaultdict(list)
+        for e in elements:
+            element_inds_from_tag[e.elem_tag].append(e.elem_ind)
+        supports = [Support.from_data(s) for s in data['supports']]
+        joints = [Joint.from_data(j) for j in data['joints']]
+        materials = [Material.from_data(m) for m in data['materials']]
+        crosssecs = [CrossSec.from_data(c) for c in data['cross_secs']]
+        # sanity checks
+        if 'node_num' in data:
+            assert len(nodes) == data['node_num']
+        if 'element_num' in data:
+            assert len(elements) == data['element_num']
+        node_id_range = list(range(len(nodes)))
+        for e in elements:
+            assert e.end_node_inds[0] in node_id_range and \
+                   e.end_node_inds[1] in node_id_range, \
+                   'element end point id not in node_list id range!'
+        num_grounded_nodes = 0
+        for n in nodes:
+            if n.is_grounded:
+                num_grounded_nodes += 1
+        # assert grounded_nodes > 0, 'The structure must have at lease one grounded node!'
+        if num_grounded_nodes == 0:
+            warnings.warn('The structure must have at lease one grounded node!')
+        if verbose:
+            print('Model: {} | Original unit: {} | Generated time: {}'.format(model_name, data['unit'], 
+                data['generate_time'] if 'generate_time' in data else ''))
+            print('Nodes: {} | Elements: {} | Supports: {} | Joints: {} | Materials: {} | Cross Secs: {} | Tag Ground: {} '.format(
+                len(nodes), len(elements), len(supports), len(joints), len(materials), len(crosssecs), num_grounded_nodes))
         return cls(nodes, elements, supports, joints, materials, crosssecs, model_name=model_name)
 
     def to_data(self):
-        from pyconmech.frame_analysis.frame_file_io import frame_to_data
-        return frame_to_data(self.nodes, self.elements, self.supports, self.joints, self.materials, self.crosssecs, self.model_name)
+        data = {'model_name' : self.model_name,
+                'unit' : self.unit,
+                'generate_time' : self.generate_time,
+                'nodes' : [n.to_data() for n in self.nodes],
+                'elements' : [e.to_data() for e in self.elements],
+                'supports' : [s.to_data() for s in self.supports.values()],
+                'joints' : [j.to_data() for j in self.joints.values()],
+                'materials' : [m.to_data() for m in self.materials.values()],
+                'cross_secs' : [cs.to_data() for cs in self.crosssecs.values()],
+                }
+        return data
 
+class LoadCase(object):
+    def __init__(self, point_loads, uniform_element_loads, gravity_load):
+        self.point_loads = point_loads
+        self.uniform_element_loads = uniform_element_loads
+        self.gravity_load = gravity_load
+
+    @classmethod
+    def from_json(cls, file_path):
+        assert os.path.exists(file_path) and "json file path does not exist!"
+        with open(file_path, 'r') as f:
+            json_data = json.loads(f.read())
+        return cls.from_data(json_data)
+    
+    @classmethod
+    def from_data(cls, data):
+        point_loads = [PointLoad.from_data(pl) for pl in data['ploads']]
+        uniform_element_loads = [UniformlyDistLoad.from_data(el) for el in data['eloads']]
+        gravity_load = None if 'gravity' not in data else data['gravity']
+        return cls(point_loads, uniform_element_loads, gravity_load)
+
+    def to_data(self):
+        data = {'ploads' : [pl.to_data() for pl in self.point_loads],
+                'eloads' : [el.to_data() for el in self.uniform_element_loads],
+                'gravity' : self.gravity_load.to_data()}
+        return data
+
+##############################################
 
 class Node(object):
     def __init__(self, point, node_ind, is_grounded):
@@ -41,7 +179,7 @@ class Node(object):
         return cls(data['point'], data['node_ind'], data['is_grounded'])
 
     def to_data(self):
-        data = {'point' : self.point, 'node_ind' : self.node_ind, 'is_grounded' : self.is_grounded}
+        data = {'point' : list(self.point), 'node_ind' : self.node_ind, 'is_grounded' : self.is_grounded}
         return data
 
     def __repr__(self):
@@ -60,7 +198,12 @@ class Element(object):
         return cls(data['end_node_inds'], data['elem_ind'], data['elem_tag'], data['bending_stiff'])
 
     def to_data(self):
-        raise NotImplementedError()
+        data = {'end_node_inds' : self.end_node_inds,
+                'elem_ind' : self.elem_ind,
+                'elem_tag' : self.elem_tag,
+                'bending_stiff' : self.bending_stiff,
+                }
+        return data
 
     def __repr__(self):
         return '{}(#{}({}),{},Bend:{})'.format(self.__class__.__name__, self.elem_ind, self.elem_tag, self.end_node_inds, self.bending_stiff)
@@ -75,7 +218,10 @@ class Support(object):
         return cls(data['condition'], data['node_ind'])
 
     def to_data(self):
-        raise NotImplementedError()
+        return {
+            'condition' : self.condition,
+            'node_ind' : self.node_ind,
+        }
 
     def __repr__(self):
         return '{}(#{},{})'.format(self.__class__.__name__, self.node_ind, self.condition)
@@ -90,7 +236,10 @@ class Joint(object):
         return cls(data['c_conditions'], data['elem_tags'])
 
     def to_data(self):
-        raise NotImplementedError()
+        return {
+            'c_conditions' : self.c_conditions,
+            'elem_tags' : self.elem_tags,
+        }
 
 # TODO: assumed to be converted to meter-based unit
 class CrossSec(object):
@@ -108,7 +257,15 @@ class CrossSec(object):
         return cls(data['A'], data['Jx'], data['Iy'], data['Iz'], data['elem_tags'], data['family'], data['name'])
 
     def to_data(self):
-        raise NotImplementedError()
+        return {
+            'A' : self.A,
+            'Jx' : self.Jx,
+            'Iy' : self.Iy,
+            'Iz' : self.Iz,
+            'elem_tags' : self.elem_tags,
+            'family' : self.family,
+            'name' : self.name,
+        }
 
     def __repr__(self):
         return 'family:{} name:{} area:{}[m2] Jx:{}[m4] Iy:{}[m4] Iz:{}[m4] applies to elements:{}'.format(
@@ -142,7 +299,17 @@ class Material(object):
         return cls(data['E'], data['G12'], data['fy'], data['density'], data['elem_tags'], data['family'], data['name'], data['type_name'])
 
     def to_data(self):
-        raise NotImplementedError()
+        return  {
+            'E' : self.E,
+            'G12' : self.G12,
+            'mu' : self.mu,
+            'fy' : self.fy,
+            'density' : self.density,
+            'elem_tags' : self.elem_tags,
+            'family' : self.family,
+            'name' : self.name,
+            'type_name' : self.type_name,
+        }
 
     def __repr__(self):
         # G3:8076[kN/cm2]
@@ -160,7 +327,11 @@ class PointLoad(object):
         return cls(data['force'], data['moment'], data['node_ind'])
 
     def to_data(self):
-        raise NotImplementedError()
+        return {
+            'force' : self.force,
+            'moment' : self.moment,
+            'node_ind' : self.node_ind,
+        }
 
     def __repr__(self):
         return 'Pointload: node_ind {} | force {} | moment {}'.format(self.node_ind, self.force, self.moment)
@@ -177,7 +348,11 @@ class UniformlyDistLoad(object):
         return cls(data['q'], data['load'], data['elem_tags'])
 
     def to_data(self):
-        raise NotImplementedError()
+        return {
+            'q' : self.q,
+            'load' : self.load,
+            'elem_tags' : self.elem_tags,
+        }
 
     def __repr__(self):
         return 'UniformlyDistLoad: element_tags {} | q {} | load {}'.format(self.elem_tags, self.q, self.load)
@@ -191,7 +366,15 @@ class GravityLoad(object):
         return cls(data['force'])
 
     def to_data(self):
-        raise NotImplementedError()
+        return {
+            'force' : self.force,
+        }
 
     def __repr__(self):
         return 'GravityLoad: {}'.format(self.force)
+
+##########################################
+
+class AnalysisResult(object):
+    def __init__(self):
+        pass